Refactored Spotlight methods, moved to view.ui package
