+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/news/NewsFragment.java
@@ -56,7 +56,7 @@ import eu.davidea.flexibleadapter.FlexibleAdapter;
 import eu.davidea.flexibleadapter.common.FlexibleItemDecoration;
 import eu.davidea.flexibleadapter.common.SmoothScrollLinearLayoutManager;
 
-import static org.pattonvillecs.pattonvilleapp.SpotlightHelper.showSpotlight;
+import static org.pattonvillecs.pattonvilleapp.view.ui.spotlight.SpotlightHelper.showSpotlight;
 
 /**
  * Fragment used within MainActivity to display the News tab
